+++ src/tree-graph.ts
@@ -1,4 +1,4 @@
-import type { Command } from "./types/command";
+import type { Command } from "./types";
 
 const branchChar = '┗ ';
 const straightChar = '┃ ';
@@ -14,36 +14,40 @@ const branchAndStraightChar = '┣ ';
 function recursiveTree(command: Command, margin: number, options: TreeGraphOptions, endedMargin: number): string {
     const marginString = '  '.repeat(endedMargin) + straightChar.repeat(margin - endedMargin);
     let result: string[] = [];
-    const subcommandsLength = (command.subcommands?.length ?? 0);
-    for (let i = 0; i < subcommandsLength; i++) {
-        const subcommand = command.subcommands![i];
-        // if it is hidden and showing them is not configured, move to the next one
-        if ((subcommand.hidden ?? false) && !options.showHidden) continue;
-        // choose right char
-        let char = branchAndStraightChar;
-        if (i === subcommandsLength - 1) {
-            char = branchChar;
-            // increase ended margin count
-            endedMargin++;
-        }
-        // get command name
-        let commandName = subcommand.name;
-        if (options.colored && subcommand.handler === null) {
-            // dim it if it does not have a handler and coloring is enabled
-            commandName = `\x1b[2m${commandName}\x1b[0m`;
-        }
-        // add it
-        result.push(
-            marginString + char + commandName
-        );
-        // if it has subcommands
-        if (subcommand.subcommands && subcommand.subcommands.length > 0) {
-            // if subcommands are hidden and showing them is not configured
-            if ((subcommand.hideSubcommands ?? false) && !options.showHiddenSubcommands) {
-                // sign that it has hidden subcommands
-                result[result.length - 1] += ' +';
-            } else {
-                result.push(recursiveTree(subcommand, margin + 1, options, endedMargin));
+    const subcommandsKeys = Object.keys(command.subcommands ?? {});
+    const subcommandsLength: number = subcommandsKeys.length;
+    if (command.subcommands) {
+        for (let i = 0; i < subcommandsLength; i++) {
+            const subcommandName = subcommandsKeys[i];
+            const subcommand = command.subcommands[subcommandName];
+            // if it is hidden and showing them is not configured, move to the next one
+            if ((subcommand.hidden ?? false) && !options.showHidden) continue;
+            // choose right char
+            let char = branchAndStraightChar;
+            if (i === subcommandsLength - 1) {
+                char = branchChar;
+                // increase ended margin count
+                endedMargin++;
+            }
+            // get command name
+            let commandName = subcommandName;
+            if (options.colored && subcommand.handler === null) {
+                // dim it if it does not have a handler and coloring is enabled
+                commandName = `\x1b[2m${commandName}\x1b[0m`;
+            }
+            // add it
+            result.push(
+                marginString + char + commandName
+            );
+            // if it has subcommands
+            if (subcommand.subcommands && Object.keys(subcommand.subcommands).length > 0) {
+                // if subcommands are hidden and showing them is not configured
+                if ((subcommand.hideSubcommands ?? false) && !options.showHiddenSubcommands) {
+                    // sign that it has hidden subcommands
+                    result[result.length - 1] += ' +';
+                } else {
+                    result.push(recursiveTree(subcommand, margin + 1, options, endedMargin));
+                }
             }
         }
     }
@@ -65,7 +69,7 @@ export type TreeGraphOptions = {
  * @options Options for the tree graph generator. Boolean values are legacy, they represent the color
  * @returns The graph
  */
-export function treeGraph(command: Command, options: TreeGraphOptions = {}): string {
+export function treeGraph(command: Command, commandName: string, options: TreeGraphOptions = {}): string {
     // legacy
     if (typeof options === 'boolean') options = { colored: options };
     // default values
@@ -74,5 +78,5 @@ export function treeGraph(command: Command, options: TreeGraphOptions = {}): str
     options.showHiddenSubcommands ??= false;
     // run
     const graph = recursiveTree(command, 0, options, 0);
-    return command.name + '\n' + graph;
+    return commandName + '\n' + graph;
 }