+++ src/types/command.ts
@@ -6,13 +6,6 @@ import { Arg } from "./arg";
  * Object type of commands
  */
 export type Command = {
-    /** Name of the command that will be used to call it.
-     * 
-     * For root command it will be used only in help
-     * and we recommend to set it to the command your CLI is called.
-     * For example `my-cli` or `node main` if it doesn't have its bin
-     */
-    name: string,
     /** Description of the command for help */
     description?: string,
     /**
@@ -31,7 +24,9 @@ export type Command = {
      * Subcommands of the command
      * For example, `my-cli command subcommand`
      */
-    subcommands?: Command[],
+    subcommands?: {
+        [key: string]: Command,
+    },
     /** 
      * Enable `--help` and `-h` options for this command?
      * `true` by default