+++ src/utils.ts
@@ -1,0 +1,7 @@
+import { treeGraph } from './tree-graph.js';
+import exitWithError from './exit-with-error.js';
+
+export {
+    exitWithError,
+    treeGraph,
+};